refactor: move type declarations to types.ts file
